fix: move audit log formatting over to the audit service, from the state resource
