+++ lib/components/services/audit/index.js
@@ -6,6 +6,7 @@ const schema = require('./schema.json')
 const generateTriggerStatement = require('./generate-trigger-statement')
 const debug = require('debug')('@wmfs/tymly-pg-plugin')
 const pgInfo = require('@wmfs/pg-info')
+const DateTime = require('luxon').DateTime
 
 class AuditService {
   async boot (options) {
@@ -90,7 +91,59 @@ class AuditService {
       }
     })
   } // loadLoads
-}
+
+  formatLogs (modelName, auditLogs) {
+    const model = this.models[modelName.replace('.', '_')]
+
+    const logs = auditLogs
+      .map(l => formatLog(l, model))
+      .reverse()
+    return logs
+  } // formatLogs
+} // AuditService
+
+function formatLog (log, model) {
+  const diffs = formatDiffs(log.diff, model)
+  const when = formatDate(log.modified)
+
+  const formatted = {
+    change: diffs.join(', \n'),
+    modified: when,
+    modifiedBy: log.modifiedBy
+  }
+  return formatted
+} // formatLog
+
+function formatDiffs (diffs, model) {
+  return Object.entries(diffs)
+    .map(([field, change]) => formatDiff(field, change, model))
+    .filter(l => l)
+} // formatDiffs
+
+function formatDiff (field, change, model) {
+  const propertyName = _.camelCase(field)
+  const property = model.properties[propertyName]
+
+  if (!property || property.audit === 'raw') {
+    return
+  }
+
+  const label = property.title || field
+
+  if (!change.from) {
+    return `${label} set to "${change.to}"`
+  }
+  if (!change.to) {
+    return `${label} "${change.from}" was cleared`
+  }
+
+  return `${label} changed from "${change.from}" to "${change.to}"`
+} // formatDiff
+
+function formatDate (jsDate) {
+  const d = DateTime.fromJSDate(jsDate)
+  return `${d.toLocaleString(DateTime.TIME_24_SIMPLE)} ${d.toLocaleString(DateTime.DATE_MED)}`
+} // formatDate
 
 module.exports = {
   schema: schema,
